requirements.txt
JPype1==1.4.1
pandas==2.1.4
openpyxl==3.1.2
tabula-py==2.9.0
pdfminer==20191125
pdfminer.six==20221105